# services/python-backend/app/api/v1/routes.py
from __future__ import annotations

import asyncio
from typing import Any, Literal, cast

from fastapi import APIRouter, Depends, Header, HTTPException, Request, status
from pydantic import ValidationError

from app.core.errors import (
    CROSS_OWNER_EVIDENCE,
    CROSS_TENANT_EVIDENCE,
    EVIDENCE_STORE_UNAVAILABLE,
    IDEMPOTENCY_IN_PROGRESS,
    IDEMPOTENCY_KEY_REUSED,
    ProviderError,
    http_status_for,
)
from app.core.idempotency import idempotency_redis_key, lock_ttl_seconds, request_hash
from app.core.metrics import IDEMPOTENCY_HITS, METRICS, PROVIDER_FAILURES, STAGE_LATENCY, TOKENS_IN, TOKENS_OUT
from app.core.security import require_service_token
from app.domain.schemas import (
    AuditRequest,
    AuditResponse,
    EvidenceIndexRequest,
    EvidenceIndexResponse,
    EvidenceMatchRequest,
    EvidenceMatchResponse,
    EvidenceSearchRequest,
    EvidenceSearchResponse,
    FinalQaRequest,
    FinalQaResponse,
    JobParseRequest,
    JobParseResponse,
    ResearchSynthesizeRequest,
    ResearchSynthesizeResponse,
    ResumeGenerateRequest,
    ResumeGenerateResponse,
    ResumeParseRequest,
    ResumeParseResponse,
    SingleRequestGenerateRequest,
    SingleRequestGenerateResponse,
)
from app.modules.audits import service as audits
from app.modules.evidence.service import index_evidence_items, search_evidence_store
from app.modules.evidence.store.factory import get_embedding_provider, get_evidence_store
from app.modules.evidence.store.protocol import EvidenceStoreError
from app.modules.guardrails.service import validate_resume_claims
from app.modules.parsing.service import parse_job_text, parse_resume_bytes
from app.modules.quality.service import authorize_final_qa_response
from app.modules.research import service as research_svc
from app.modules.retrieval import service as retrieval
from app.providers.factory import get_provider

router = APIRouter(prefix="/v1", dependencies=[Depends(require_service_token)])


def _raise_provider(exc: Exception) -> None:
    if isinstance(exc, ProviderError):
        raise HTTPException(status_code=http_status_for(exc.code), detail={"code": exc.code, "message": exc.message}) from exc
    if isinstance(exc, RuntimeError):
        code = str(exc).split(":", 1)[0]
        raise HTTPException(status_code=http_status_for(code), detail={"code": code, "message": str(exc)}) from exc
    if isinstance(exc, ValidationError):
        raise HTTPException(
            status_code=422,
            detail={"code": "PROVIDER_OUTPUT_INVALID", "message": str(exc)},
        ) from exc
    if isinstance(exc, ValueError) and str(exc).startswith("GUARDRAIL_VIOLATION"):
        raise HTTPException(
            status_code=422,
            detail={"code": "GUARDRAIL_VIOLATION", "message": str(exc)},
        ) from exc
    if isinstance(exc, ValueError) and str(exc).startswith("REFINEMENT_NOT_APPLICABLE"):
        raise HTTPException(
            status_code=422,
            detail={"code": "REFINEMENT_NOT_APPLICABLE", "message": str(exc)},
        ) from exc
    raise exc


def _assert_evidence_scope(tenant_id: str, user_id: str, evidence: list[Any]) -> None:
    for item in evidence:
        if item.tenant_id != tenant_id:
            raise HTTPException(
                status_code=403,
                detail={"code": CROSS_TENANT_EVIDENCE, "message": "Evidence tenant mismatch"},
            )
        if item.owner_user_id != user_id:
            raise HTTPException(
                status_code=403,
                detail={"code": CROSS_OWNER_EVIDENCE, "message": "Evidence owner mismatch"},
            )


async def _with_idempotency(
    request: Request,
    *,
    operation: str,
    tenant_id: str,
    user_id: str,
    idempotency_key: str | None,
    body_dict: dict[str, Any],
    handler: Any,
) -> Any:
    if not idempotency_key:
        return await handler()

    store = request.app.state.idempotency
    settings = request.app.state.settings
    key = idempotency_redis_key(tenant_id, user_id, operation, idempotency_key)
    digest = request_hash(body_dict)
    lock_ttl = lock_ttl_seconds(
        response_ttl_seconds=settings.idempotency_ttl_seconds,
        lock_budget_seconds=settings.idempotency_lock_ttl_seconds,
    )
    try:
        begun = await store.begin(key, digest, lock_ttl)
    except ProviderError as exc:
        if exc.code in {IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_IN_PROGRESS}:
            raise HTTPException(
                status_code=409,
                detail={"code": exc.code, "message": exc.message},
            ) from exc
        raise HTTPException(status_code=http_status_for(exc.code), detail={"code": exc.code, "message": exc.message}) from exc

    if begun.cached_response is not None:
        METRICS.incr(IDEMPOTENCY_HITS)
        return begun.cached_response

    owner = begun.owner
    if owner is None:
        raise HTTPException(
            status_code=500,
            detail={"code": "IDEMPOTENCY_LOCK_MISSING", "message": "Idempotency lock acquired without owner"},
        )

    # Heartbeat renews lock every lock_ttl/3 while the provider call runs.
    stop_heartbeat = asyncio.Event()

    async def _heartbeat() -> None:
        interval = max(1.0, lock_ttl / 3)
        while not stop_heartbeat.is_set():
            try:
                await asyncio.wait_for(stop_heartbeat.wait(), timeout=interval)
                return
            except TimeoutError:
                renewed = await store.renew(key, owner, lock_ttl)
                if not renewed:
                    return

    heartbeat_task = asyncio.create_task(_heartbeat())
    try:
        # Renew once up-front with full lock TTL before the provider call.
        await store.renew(key, owner, lock_ttl)
        with METRICS.time_block(f"{STAGE_LATENCY}.{operation}"):
            result = await handler()
        payload = result.model_dump(mode="json") if hasattr(result, "model_dump") else result
        usage = getattr(result, "usage", None)
        if usage is not None:
            if getattr(usage, "input_tokens", None):
                METRICS.incr(TOKENS_IN, int(usage.input_tokens or 0))
            if getattr(usage, "output_tokens", None):
                METRICS.incr(TOKENS_OUT, int(usage.output_tokens or 0))
        await store.complete(key, digest, payload, settings.idempotency_ttl_seconds, owner)
        return result
    except Exception:
        METRICS.incr(PROVIDER_FAILURES)
        await store.release(key, owner)
        raise
    finally:
        stop_heartbeat.set()
        heartbeat_task.cancel()
        try:
            await heartbeat_task
        except asyncio.CancelledError:
            pass


@router.post("/resumes/parse", response_model=ResumeParseResponse)
async def resumes_parse(body: ResumeParseRequest) -> ResumeParseResponse:
    try:
        return await parse_resume_bytes(body.filename, body.content_type, body.content_base64)
    except ValueError as exc:
        code = str(exc)
        messages = {
            "IMAGE_ONLY_PDF_OCR_REQUIRED": "This PDF appears to contain scanned images. Upload a text-based PDF or DOCX, or enter your details manually. OCR is not available.",
            "PDF_ENCRYPTED": "This PDF is password-protected. Upload an unlocked PDF or DOCX.",
            "CORRUPT_PDF": "This PDF could not be read. Try exporting again or upload a DOCX.",
            "EMPTY_DOCUMENT": "This file appears empty. Upload a text-based PDF or DOCX.",
            "LEGACY_DOC_UNSUPPORTED": "Legacy .doc Word files are not supported. Open the file in Word or LibreOffice and save it as .docx, then upload again.",
            "DOCUMENT_TOO_LARGE": "File is too large. Maximum size is 10 MB.",
            "PDF_PAGE_LIMIT_EXCEEDED": "This PDF exceeds the 30-page limit.",
            "PARSE_TIMEOUT": "Parsing timed out. Try a smaller file or enter details manually.",
            "INVALID_PDF_MAGIC": "This file is not a valid PDF.",
            "INVALID_DOCX_MAGIC": "This file is not a valid DOCX.",
            "UNSUPPORTED_DOCUMENT_TYPE": "Unsupported file type. Upload a PDF or DOCX.",
            "PYTHON_BACKEND_UNAVAILABLE": "Resume parsing is temporarily unavailable.",
        }
        raise HTTPException(
            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
            detail={"code": code, "message": messages.get(code, "Unsupported or unreadable document")},
        ) from exc


@router.post("/jobs/parse", response_model=JobParseResponse)
async def jobs_parse(body: JobParseRequest) -> JobParseResponse:
    parsed = parse_job_text(body.job_text, body.company, body.role)
    return JobParseResponse(**parsed)


async def _research_handler(request: Request, body: ResearchSynthesizeRequest) -> ResearchSynthesizeResponse:
    try:
        provider = get_provider("generation", request)
        result, latency, usage = await provider.synthesize_research(
            company=body.company,
            role=body.role,
            sources=body.sources,
            job_description=body.job_description,
            team=body.team, product=body.product, business_unit=body.business_unit,
        )
    except Exception as exc:
        _raise_provider(exc)
        raise
    typed = result if isinstance(result, ResearchSynthesizeResponse) else ResearchSynthesizeResponse.model_validate(result)
    claim_leaks = research_svc.research_must_not_become_claims(typed.findings)
    if claim_leaks:
        raise HTTPException(
            status_code=422,
            detail={"code": "RESEARCH_AS_CANDIDATE_CLAIM", "message": "Research must not become candidate claims"},
        )
    return typed.model_copy(
        update={
            "provider": usage.provider,
            "model": usage.model,
            "latency_ms": latency,
            "usage": usage,
        }
    )


@router.post("/research/synthesize", response_model=ResearchSynthesizeResponse)
async def research_synthesize(request: Request, body: ResearchSynthesizeRequest,
                              idempotency_key: str | None = Header(default=None)) -> Any:
    return await _with_idempotency(request, operation="research", tenant_id=body.context.tenant_id,
        user_id=body.context.user_id, idempotency_key=idempotency_key, body_dict=body.model_dump(mode="json"),
        handler=lambda: _research_handler(request, body))


def _store_backend(request: Request) -> Literal["memory", "postgres"]:
    settings = request.app.state.settings
    return cast(Literal["memory", "postgres"], settings.evidence_store_backend())


@router.post("/evidence/index", response_model=EvidenceIndexResponse)
async def evidence_index(request: Request, body: EvidenceIndexRequest) -> EvidenceIndexResponse:
    """Index evidence via EvidenceStore. Production uses postgres; demo may use memory.

    Process-global `_INDEX` is never authoritative in production (no silent memory fallback).
    """
    settings = request.app.state.settings
    backend = settings.evidence_store_backend()
    if settings.app_mode == "production" and backend != "postgres":
        raise HTTPException(
            status_code=503,
            detail={
                "code": EVIDENCE_STORE_UNAVAILABLE,
                "message": "Production requires postgres evidence store",
            },
        )
    _assert_evidence_scope(body.context.tenant_id, body.context.user_id, body.evidence)
    try:
        store = get_evidence_store()
        embedder = get_embedding_provider()
        indexed = await index_evidence_items(
            store,
            embedder,
            tenant_id=body.context.tenant_id,
            owner_user_id=body.context.user_id,
            evidence=body.evidence,
        )
    except EvidenceStoreError as exc:
        raise HTTPException(
            status_code=http_status_for(exc.code),
            detail={"code": exc.code, "message": exc.message},
        ) from exc
    # Keep demo process-local cache in sync for legacy unit tests only (never in production)
    if settings.app_mode != "production" and backend == "memory":
        retrieval.index_evidence(body.context.tenant_id, body.context.user_id, body.evidence)
    return EvidenceIndexResponse(
        indexed=indexed,
        tenant_id=body.context.tenant_id,
        owner_user_id=body.context.user_id,
        experimental=backend == "memory",
        store_backend=backend,
    )


@router.post("/evidence/search", response_model=EvidenceSearchResponse)
async def evidence_search(request: Request, body: EvidenceSearchRequest) -> EvidenceSearchResponse:
    settings = request.app.state.settings
    backend = settings.evidence_store_backend()
    if settings.app_mode == "production" and backend != "postgres":
        raise HTTPException(
            status_code=503,
            detail={
                "code": EVIDENCE_STORE_UNAVAILABLE,
                "message": "Production requires postgres evidence store",
            },
        )
    if body.owner_user_id != body.context.user_id:
        raise HTTPException(status_code=403, detail={"code": "CROSS_OWNER_SEARCH", "message": "Cannot search another candidate"})
    try:
        store = get_evidence_store()
        embedder = get_embedding_provider()
        hits = await search_evidence_store(
            store,
            embedder,
            tenant_id=body.context.tenant_id,
            owner_user_id=body.owner_user_id,
            query=body.query,
            limit=body.limit,
        )
    except EvidenceStoreError as exc:
        raise HTTPException(
            status_code=http_status_for(exc.code),
            detail={"code": exc.code, "message": exc.message},
        ) from exc
    return EvidenceSearchResponse(
        hits=hits,
        experimental=backend == "memory",
        store_backend=backend,
    )


async def _evidence_match_handler(request: Request, body: EvidenceMatchRequest) -> EvidenceMatchResponse:
    """Authoritative request-scoped match. Optionally enrich notes via store search (non-authoritative)."""
    _assert_evidence_scope(body.context.tenant_id, body.context.user_id, body.evidence)
    try:
        provider = get_provider("generation", request)
        result, latency, usage = await provider.match_evidence(
            requirements=body.requirements,
            evidence=body.evidence,
            research_findings=body.research_findings,
            job_description=body.job_description,
            role=body.role,
            company=body.company,
        )
    except Exception as exc:
        _raise_provider(exc)
        raise
    typed = result if isinstance(result, EvidenceMatchResponse) else EvidenceMatchResponse.model_validate(result)
    return typed.model_copy(
        update={
            "provider": usage.provider,
            "model": usage.model,
            "latency_ms": latency,
            "usage": usage,
        }
    )


@router.post("/evidence/match", response_model=EvidenceMatchResponse)
async def evidence_match(request: Request, body: EvidenceMatchRequest,
                         idempotency_key: str | None = Header(default=None)) -> Any:
    _assert_evidence_scope(body.context.tenant_id, body.context.user_id, body.evidence)
    return await _with_idempotency(request, operation="resume-plan", tenant_id=body.context.tenant_id,
        user_id=body.context.user_id, idempotency_key=idempotency_key, body_dict=body.model_dump(mode="json"),
        handler=lambda: _evidence_match_handler(request, body))


async def _generate_handler(request: Request, body: ResumeGenerateRequest) -> ResumeGenerateResponse:
    _assert_evidence_scope(body.context.tenant_id, body.context.user_id, body.evidence)
    _ = body.research_findings  # untrusted — questions only via generation path
    absolute = body.absolute_version if body.absolute_version is not None else 0
    cycle = body.cycle_step if body.cycle_step is not None else absolute % 5
    try:
        provider = get_provider("generation", request)
        resume, latency, usage = await provider.generate_resume(
            absolute_version=absolute,
            cycle_step=cycle,
            version_number=absolute,
            evidence=body.evidence,
            allowed_technologies=body.allowed_technologies,
            job_description=body.job_description,
            job_requirements=body.job_requirements,
            previous_resume=body.previous_resume,
            accepted_findings=body.accepted_findings,
            rejected_findings=body.rejected_findings,
            mistake_memory=body.mistake_memory,
            research_findings=body.research_findings,
            user_confirmations=body.user_confirmations,
            refinement_instruction=body.refinement_instruction,
            evidence_matches=body.evidence_matches,
            resume_plan=body.resume_plan,
            final_qa_repair=body.final_qa_repair,
        )
    except Exception as exc:
        _raise_provider(exc)
        raise
    violations = validate_resume_claims(
        resume,
        body.evidence,
        body.allowed_technologies,
        tenant_id=body.context.tenant_id,
        owner_user_id=body.context.user_id,
        job_description=body.job_description,
        research_findings=body.research_findings,
        user_confirmations=body.user_confirmations,
    )
    if violations:
        raise HTTPException(
            status_code=422,
            detail={
                "code": "GUARDRAIL_VIOLATION",
                "message": "Resume failed claim checks",
                "details": {"violations": violations},
            },
        )
    return ResumeGenerateResponse(
        resume=resume,
        provider=usage.provider if usage else provider.name,
        model=usage.model if usage else provider.model,
        prompt_version=usage.prompt_version if usage else "resume-generation@python-v2",
        latency_ms=latency,
        usage=usage,
    )


@router.post("/resumes/generate", response_model=ResumeGenerateResponse)
async def resumes_generate(
    request: Request,
    body: ResumeGenerateRequest,
    idempotency_key: str | None = Header(default=None, alias="Idempotency-Key"),
) -> ResumeGenerateResponse | dict[str, Any]:
    return cast(
        ResumeGenerateResponse | dict[str, Any],
        await _with_idempotency(
            request,
            operation="generate",
            tenant_id=body.context.tenant_id,
            user_id=body.context.user_id,
            idempotency_key=idempotency_key,
            body_dict=body.model_dump(mode="json"),
            handler=lambda: _generate_handler(request, body),
        ),
    )


@router.post("/resumes/audit", response_model=AuditResponse)
async def resumes_audit(
    request: Request,
    body: AuditRequest,
    idempotency_key: str | None = Header(default=None, alias="Idempotency-Key"),
) -> AuditResponse | dict[str, Any]:
    _assert_evidence_scope(body.context.tenant_id, body.context.user_id, body.evidence)

    async def _handler() -> AuditResponse:
        try:
            provider = get_provider(audits.lens_to_role(body.lens), request)
            result, _, usage = await provider.audit(
                lens=body.lens,
                reviews_version=body.reviews_version,
                produces_version=body.produces_version,
                resume=body.resume,
                evidence=body.evidence,
                job_description=body.job_description,
                allowed_technologies=body.allowed_technologies,
                tenant_id=body.context.tenant_id,
                owner_user_id=body.context.user_id,
            )
        except Exception as exc:
            _raise_provider(exc)
            raise
        typed = result if isinstance(result, AuditResponse) else AuditResponse.model_validate(result)
        if typed.usage is None:
            typed = typed.model_copy(update={"usage": usage})
        if not typed.rejected_findings and typed.findings:
            accepted, rejected = audits.adjudicate_findings(
                typed.findings,
                body.evidence,
                body.allowed_technologies,
                tenant_id=body.context.tenant_id,
                owner_user_id=body.context.user_id,
            )
            typed = typed.model_copy(update={"findings": accepted, "rejected_findings": rejected})
        return typed

    return cast(
        AuditResponse | dict[str, Any],
        await _with_idempotency(
            request,
            operation="audit",
            tenant_id=body.context.tenant_id,
            user_id=body.context.user_id,
            idempotency_key=idempotency_key,
            body_dict=body.model_dump(mode="json"),
            handler=_handler,
        ),
    )


@router.post("/resumes/regenerate", response_model=ResumeGenerateResponse)
async def resumes_regenerate(
    request: Request,
    body: ResumeGenerateRequest,
    idempotency_key: str | None = Header(default=None, alias="Idempotency-Key"),
) -> ResumeGenerateResponse | dict[str, Any]:
    return cast(
        ResumeGenerateResponse | dict[str, Any],
        await _with_idempotency(
            request,
            operation="regenerate",
            tenant_id=body.context.tenant_id,
            user_id=body.context.user_id,
            idempotency_key=idempotency_key,
            body_dict=body.model_dump(mode="json"),
            handler=lambda: _generate_handler(request, body),
        ),
    )


@router.post("/resumes/final-qa", response_model=FinalQaResponse)
async def resumes_final_qa(
    request: Request,
    body: FinalQaRequest,
    idempotency_key: str | None = Header(default=None, alias="Idempotency-Key"),
) -> FinalQaResponse | dict[str, Any]:
    _assert_evidence_scope(body.context.tenant_id, body.context.user_id, body.evidence)

    async def _handler() -> FinalQaResponse:
        try:
            provider = get_provider("final-review", request)
            result, _, usage = await provider.final_qa(
                resume=body.resume,
                evidence=body.evidence,
                deterministic_checks=body.deterministic_checks,
                allowed_technologies=body.allowed_technologies,
            )
            typed = authorize_final_qa_response(
                result,
                resume=body.resume,
                evidence=body.evidence,
                allowed_technologies=body.allowed_technologies,
                deterministic_checks=body.deterministic_checks,
            )
        except Exception as exc:
            _raise_provider(exc)
            raise
        if typed.usage is None:
            typed = typed.model_copy(update={"usage": usage})
        provider_name = usage.provider if usage else typed.provider
        model_name = usage.model if usage else typed.model
        typed = typed.model_copy(update={
            "provider": provider_name,
            "model": model_name,
        })
        return typed

    return cast(
        FinalQaResponse | dict[str, Any],
        await _with_idempotency(
            request,
            operation="final-qa",
            tenant_id=body.context.tenant_id,
            user_id=body.context.user_id,
            idempotency_key=idempotency_key,
            body_dict=body.model_dump(mode="json"),
            handler=_handler,
        ),
    )


@router.post("/resumes/generate-once", response_model=SingleRequestGenerateResponse)
async def resumes_generate_once(request: Request, body: SingleRequestGenerateRequest) -> dict[str, Any]:
    from app.modules.single_request import generate_once
    _assert_evidence_scope(body.context.tenant_id, body.context.user_id, body.evidence)
    if body.previous_resume or body.refinement_instruction or body.final_qa_repair or body.absolute_version != 0:
        raise HTTPException(422, detail={"code": "INITIAL_GENERATION_ONLY"})
    try:
        return await generate_once(request, body)
    except Exception as exc:
        _raise_provider(exc)
        raise